+++ src/dialogs/linkdialog.cpp
@@ -21,6 +21,7 @@
 
 #include "entities/notefolder.h"
 #include "mainwindow.h"
+#include "services/scriptingservice.h"
 #include "services/settingsservice.h"
 #include "ui_linkdialog.h"
 #include "widgets/navigationwidget.h"
@@ -225,6 +226,8 @@ QString LinkDialog::getURL() const {
 
     return url;
 }
+
+void LinkDialog::setURL(const QString &text) { ui->urlEdit->setText(text); }
 
 QString LinkDialog::getLinkName() const { return ui->nameLineEdit->text().trimmed(); }
 
@@ -571,6 +574,14 @@ void LinkDialog::on_tabWidget_currentChanged(int index) {
 }
 
 void LinkDialog::startTitleFetchRequest(const QUrl &url) {
+    const QString title =
+        ScriptingService::instance()->callFetchUrlTitleHook(url.toString(QUrl::FullyEncoded));
+
+    if (!title.isEmpty()) {
+        setLinkName(title);
+        return;
+    }
+
     ui->downloadProgressBar->show();
     QNetworkRequest networkRequest(url);